fix loading spinner
